python/requirements.txt
python-dotenv
google-auth
requests
